+++ EMIExpert/src/com/emiexpert/Utility.java
@@ -4,8 +4,8 @@ package com.emiexpert;
 
 public class Utility {
 	
-	public static double getRoundOfDouble(double value){
-		double roundValue = Math.round( value * 100.0 ) / 100.0;
+	public static long getRoundOfDouble(double value){
+		long roundValue = Math.round( value * 100 ) / 100;
 		return roundValue;
 		
 		/*DecimalFormat df=new DecimalFormat("0.00");
